:bulb: wait exemple
